Fix FRC service and update Battery service
Change battery service to take a uint8_t as this is the same
interface as e.g. SHT40 Gadgets use. Since battery can only be
between 0% and 100% a unit8_t is enough.

Use copy for FRC callback instead of reference in the lambda
function, since else the function pointer is not valid anymore.

+++ src/BatteryBleService.h
@@ -11,7 +11,7 @@ public:
       : IBleServiceProvider(bleLibrary) {}
   bool begin() override;
 
-  void setBatteryLevel(int value) const;
+  void setBatteryLevel(uint8_t value) const;
 };
 
 #endif // ARDUINO_UPT_BLE_SERVER_BATTERY_BLE_SERVICE_H
